New version

+++ Source/gitpush.ps1
@@ -6,10 +6,10 @@ param (
 	[string]$comment = "New version"
  )
 
- if (Test-Path $zipname) { rm $zipname; }
+if (Test-Path $zipname) { rm $zipname; }
 # Clean up binary folder since we will be generating new binaries
 rm ..\Binary\*.* -Force -Recurse
-rm obj -Force -Recurse
+if (Test-Path obj) { rm obj -Force -Recurse }
 # build new version
 msbuild /verbosity:minimal $solution
 # Compress bin\debug and move to binary folder
